Refactor blanket testing utility into new file
We're going to need to use the function in another test file.

Also, move the blanket created by the utility down so that it crosses
the x-axis (i.e., it has area either side of z=0). This means we don't
get NoneType issues when calling `slice_shape` in
BlanketBuilder.build_xy using the xy-plane. Basically, the blanket must
cross z=0 for the xy part of the build to work correctly.
